Made behavior tree dynamically reconfigurable
I'm not sure if it's a good approach to clone the entire tree every time
to instantiate a new AI.
Creating factory class for each of the node types can be more explicit
and may be able to avoid undesired transferring of member variable states,
but it would be too verbose and tedious to repeat class definition twice.
We gonna have a lot more types of node, after all.